SCardGetStatusChange(): add a cast to fix a compiler warning
winscard_clnt.c:1890: warning: comparison between signed and unsigned
integer expressions


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@5797 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/winscard_clnt.c
@@ -1887,7 +1887,7 @@ LONG SCardGetStatusChange(SCARDCONTEXT hContext, DWORD dwTimeout,
 			currentReaderCount++;
 
 	/* catch possible sign extension problems from 32 to 64-bits integers */
-	if (-1 == dwTimeout)
+	if ((DWORD)-1 == dwTimeout)
 		dwTimeout = INFINITE;
 	if (INFINITE == dwTimeout)
 		dwTime = 60*1000;	/* "infinite" timeout */
